requirements.txt
# Core dependencies
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3

# Database
psycopg2-binary==2.9.6

# Machine Learning
scikit-learn==1.3.0

# API
fastapi==0.95.2
uvicorn==0.22.0

# Development
black==23.3.0
flake8==6.0.0
pytest==7.3.1
